+++ changelog.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Improved getBinDirectory performance
+
 ## [3.0.3]
 
 - Fixed globally installed packages incorrectly being picked up